+++ NEWS.md
@@ -1,5 +1,7 @@
 # lobstr 1.1.1
 
+* Fix PROTECT error.
+
 * Remove UTF-8 charatecrs from comments
 
 # lobstr 1.1.0